refactor(core): merge CoreIsolate and EsIsolate (#7370)
