+++ cli/main.rs
@@ -79,9 +79,9 @@ use crate::permissions::Permissions;
 use crate::tsc::TargetLib;
 use crate::worker::MainWorker;
 use deno_core::v8_set_flags;
+use deno_core::CoreIsolate;
 use deno_core::Deps;
 use deno_core::ErrBox;
-use deno_core::EsIsolate;
 use deno_core::ModuleSpecifier;
 use deno_doc as doc;
 use deno_doc::parser::DocFileLoader;
@@ -217,7 +217,7 @@ async fn print_file_info(
   {
     output.map = source_map.filename.to_str().map(|s| s.to_owned());
   }
-  let es_state_rc = EsIsolate::state(&worker.isolate);
+  let es_state_rc = CoreIsolate::state(&worker.isolate);
   let es_state = es_state_rc.borrow();
 
   if let Some(deps) = es_state.modules.deps(&module_specifier) {
