+++ core/lib.rs
@@ -12,7 +12,6 @@ mod basic_state;
 mod bindings;
 mod core_isolate;
 mod errors;
-mod es_isolate;
 mod flags;
 mod module_specifier;
 mod modules;
@@ -37,8 +36,6 @@ pub use crate::core_isolate::StartupData;
 pub use crate::errors::AnyError;
 pub use crate::errors::ErrBox;
 pub use crate::errors::JSError;
-pub use crate::es_isolate::EsIsolate;
-pub use crate::es_isolate::EsIsolateState;
 pub use crate::flags::v8_set_flags;
 pub use crate::module_specifier::ModuleResolutionError;
 pub use crate::module_specifier::ModuleSpecifier;
